remove HideKeyboardFormAccessoryBar preference setting
NOTE, this setting likely wasn't being set anyway. The form accessory bar was showing up in the app.
I believe the issue was/is, the second argument is being passed a boolean, when it may need to be a string
(am testing this). The Meteor guide shows conflicting information. The example shows passing a boolean, the
function signature uses a string.